Variants accumulate text from ancestors rather than "overwriting"
So, if you have three monsters, A, B, and C, and C is a variant of B which
is a variant of A, the displayed description and notes when looking at
C will be A's value + B's value + C's value, rather than just C's value.
Variants no longer copy description and notes.